requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
instasent==0.1.4
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
MouseInfo==0.1.3
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
soupsieve==2.3.1
sqlparse==0.4.2
twilio==7.7.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
wikipedia==1.4.0
zipp==3.7.0
